requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.49
numpy==1.24.3
faiss-cpu==1.7.4
tensorflow==2.10.1
whoosh==2.7.4